Updated Mapview to a different nav item
